+++ db/query/players.sql
@@ -5,10 +5,11 @@ SELECT * FROM players;
 SELECT * FROM players WHERE player_id = $1;
 
 -- name: CreatePlayer :one
-INSERT INTO players (name, rating) VALUES ($1, $2) RETURNING *;
+INSERT INTO players (name, rating, image_url) VALUES ($1, $2, $3) RETURNING *;
 
 -- name: UpdatePlayer :one
-UPDATE players SET name = $2, rating = $3 WHERE player_id = $1 RETURNING *;
+UPDATE players SET name = $2, rating = $3, image_url = $4 WHERE player_id = $1 RETURNING *;
 
 -- name: DeletePlayer :one
 DELETE FROM players WHERE player_id = $1 RETURNING *;
+