+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 1.0.0-a0 (2024-04-15)
+
+### BREAKING CHANGE
+
+- need database migration with breaking tables
+
+### 🔐🚧📈✏️ 💩👽️🍻💬🥚🌱🚩🥅🩺 Others
+
+- **services**: generate DI for services
+
+### 🗃️ Database
+
+- **sqlc**: create queries and add players table
+
+### 🚨 Linting
+
+- fix lint
+
 ## 0.0.2 (2024-04-12)
 
 ### 🐛🚑️ Fixes